refactor: initial refactoring of codes for modal
