+++ src/components/DailyCalorieIntake/DailyCalorieIntake.module.scss
@@ -81,3 +81,72 @@
   letter-spacing: 0.04em;
   color: var(--second-text-color);
 }
+
+.modalContent {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.modalTitle {
+  width: 75%;
+  font-family: "Verdana", sans-serif;
+  line-height: 140%;
+  font-weight: 700;
+  font-size: 26px;
+  text-align: center;
+  color: #212121;
+}
+
+.modalCalories {
+  margin: 20px 0 32px;
+  font-family: "Verdana", sans-serif;
+  font-weight: 700;
+  font-size: 48px;
+  letter-spacing: 0.04em;
+  text-align: center;
+  color: #264061;
+}
+
+.modalFoodContainer {
+  width: 330px;
+}
+
+.divider {
+  color: #E0E0E0;
+}
+
+.modalFood {
+  font-family: "Verdana", sans-serif;
+  font-weight: 700;
+  font-size: 14px;
+  letter-spacing: 0.04em;
+  color: #212121;
+}
+
+.modalFoodSearch {
+  outline: none;
+  border-radius: 10px;
+  width: 350px;
+  padding-left: 5px;
+  margin-top: 10px;
+}
+
+.modalFoodList {
+  display: flex;
+  flex-direction: column;
+  gap: 14px;
+  margin: 25px 0;
+  padding: 20px 0 40px 40px;
+  height: 150px;
+  width: 350px;
+  overflow-y: auto;
+}
+
+.modalFoodListItem {
+  list-style-type: decimal;
+  font-family: "Verdana", sans-serif;
+  font-size: 14px;
+  letter-spacing: 0.04em;
+  color: #9b9faa;
+}
