+++ src/components/Modal/Modal.module.scss
@@ -103,59 +103,3 @@
   }
 }
 
-.modalContent {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.modalTitle {
-  width: 75%;
-  font-family: "Verdana", sans-serif;
-  line-height: 140%;
-  font-weight: 700;
-  font-size: 26px;
-  text-align: center;
-  color: #212121;
-}
-
-.modalCalories {
-  margin: 20px 0 32px;
-  font-family: "Verdana", sans-serif;
-  font-weight: 700;
-  font-size: 48px;
-  letter-spacing: 0.04em;
-  text-align: center;
-  color: #264061;
-}
-
-.modalFoodContainer {
-  width: 330px;
-}
-
-.divider {
-  color: #E0E0E0;
-}
-
-.modalFood {
-  font-family: "Verdana", sans-serif;
-  font-weight: 700;
-  font-size: 14px;
-  letter-spacing: 0.04em;
-  color: #212121;
-}
-
-.modalFoodList {
-  display: flex;
-  flex-direction: column;
-  gap: 14px;
-  padding: 20px 0 40px 20px;
-}
-
-.modalFoodListItem {
-  list-style-type: decimal;
-  font-family: "Verdana", sans-serif;
-  font-size: 14px;
-  letter-spacing: 0.04em;
-  color: #9b9faa;
-}
